Reduce STAL timeout to 4 minutes
git-svn-id: https://joinup.ec.europa.eu/svn/mocca/trunk@1246 8a26b1a7-26f0-462f-b9ef-d0e30c41f5a4

+++ BKUOnline/src/main/java/at/gv/egiz/stal/service/impl/RequestBrokerSTALFactory.java
@@ -49,8 +49,8 @@ public class RequestBrokerSTALFactory implements STALFactory, ComponentMXBean {
 
     public static final String APPLET_TIMEOUT = "AppletTimeout";
 
-    public int getAppletTimeout() {
-      return configuration.getInteger(APPLET_TIMEOUT, -1);
+    public long getAppletTimeout() {
+      return configuration.getLong(APPLET_TIMEOUT, -1);
     }
 
   }
